+++ Tower-Souls/Source.cpp
@@ -1,15 +1,18 @@
 #include "Game.h"
+#include "XML.h"
 
 int main()
 {
 
-	/*Stage *stage = new Stage(grid);
+	//Stage *stage = new Stage(grid);
 	Player *p = new Player();
-	std::vector<Stage*> st = { stage };
-	Tower *t = new Tower(st);
-	Game *g = new Game(p, t);
+	//std::vector<Stage*> st = { stage };
+	XML *x = new XML();
+	Tower tower = x->load();
+	//Tower *t = new Tower();
+	Game *g = new Game(p, tower);
 
 	g->play();
-	*/
+	
 	return 0;
 }
